renaming firstapp->api_server and sba_app->web_app
