+++ django_tutorial/first_project/api_server/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class APIServerConfig(AppConfig):
+    name = 'api_server'
