+++ django_tutorial/first_project/api_server/templates/api_server/index.html
@@ -1,4 +1,4 @@
-{% extends 'layout/layout1.html' %}
+{% extends 'api_server/layout/layout1.html' %}
 
 {% block title %}Title page{% endblock %}
