--- django_tutorial/first_project/firstapp/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class FirstappConfig(AppConfig):
-    name = 'firstapp'
