--- django_tutorial/first_project/firstapp/urls.py
@@ -1,8 +1,0 @@
-from django.conf.urls import url
-from firstapp import views
-
-app_name = 'firstapp'
-urlpatterns = [
-    url(r'^$', views.index, name='index'),
-    url(r'^api/v1/regions$', views.regions, name='regions'),
-]
