--- django_tutorial/first_project/sba_app/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class SbaAppConfig(AppConfig):
-    name = 'sba_app'
