--- django_tutorial/first_project/sba_app/static/sba_app/SBA_Loans.html
@@ -1,7 +1,0 @@
-<html>
-<header>
-</header>
-<body>
-<script type='text/javascript' src='https://qa.visualizedata.ucop.edu/javascripts/api/viz_v1.js'></script><div class='tableauPlaceholder' style='width: 1353px; height: 560px;'><object class='tableauViz' width='1353' height='560' style='display:none;'><param name='host_url' value='https%3A%2F%2Fqa.visualizedata.ucop.edu%2F' /> <param name='site_root' value='&#47;t&#47;ITS' /><param name='name' value='SBALoans&#47;ZIP_Analysis' /><param name='tabs' value='yes' /><param name='toolbar' value='yes' /><param name='showAppBanner' value='false' /><param name='showShareOptions' value='true' /></object></div>
-</body>
-</html>
