+++ django_tutorial/first_project/web_app/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class WebAppConfig(AppConfig):
+    name = 'web_app'
