+++ django_tutorial/first_project/web_app/static/web_app/js/index.js
