+++ django_tutorial/first_project/web_app/tests.py
