+++ django_tutorial/first_project/web_app/views.py
@@ -4,4 +4,4 @@ from django.views.decorators.http import require_GET
 @require_GET
 def index(request):
     """base page for the front end app -- just static HTML, rest is rendered by JS"""
-    return render(request, 'index.html')
+    return render(request, 'web_app/index.html')
